Separated periodic table functionality from Main.
